+++ src/main/java/com/example/Projeto/student/config/StudentConfig.java
@@ -18,7 +18,9 @@ public class StudentConfig {
 
             Student rennan = new Student("Rennan", "rennan@gmail.com", LocalDate.of(2006, Month.FEBRUARY, 18));
 
-            studentRepository.saveAll(List.of(yuri,rennan));
+            Student erlon = new Student("Erlon", "erlon@gmail.com", LocalDate.of(2006, Month.FEBRUARY, 18));
+
+            studentRepository.saveAll(List.of(yuri,rennan, erlon));
         };
     }
 }
