+++ src/main/java/com/example/Projeto/student/controller/DeleteStudentsController.java
@@ -9,7 +9,7 @@ import java.util.List;
 
 @RestController
 @RequestMapping("api/v1/students")
-public class StudentController {
+public class ListStudentController {
     private final StudentService studentService;
 
     @Autowired
@@ -17,26 +17,9 @@ public class StudentController {
         this.studentService = studentService;
     }
 
-    @GetMapping()
-    public List<Student> getStudents(){
-        return studentService.getStudents();
-    }
-
-    @PostMapping()
-    public void insertStudents(@RequestBody Student student){
-        studentService.addStudents(student);
-    }
-
     @DeleteMapping(path = "{studentId}")
     public void deleteStudents(@PathVariable("studentId") Long studentId){
         studentService.deleteStudents(studentId);
     }
 
-    @PutMapping(path = "{studentId}")
-    public void updateStudent(@PathVariable("studentId") Long studentId,
-                              @RequestParam(required = false) String name,
-                              @RequestParam(required = false) String email){
-        studentService.updateStudent(studentId, name, email);
-
-    }
 }
