+++ src/main/java/com/example/Projeto/student/controller/GetStudentsController.java
@@ -1,0 +1,25 @@
+package com.example.Projeto.student;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.*;
+
+import java.time.LocalDate;
+import java.time.Month;
+import java.util.List;
+
+@RestController
+@RequestMapping("api/v1/students")
+public class ListStudentController {
+    private final StudentService studentService;
+
+    @Autowired
+    public StudentController(StudentService studentService) {
+        this.studentService = studentService;
+    }
+
+    @GetMapping()
+    public List<Student> getStudents(){
+        return studentService.getStudents();
+    }
+
+}
