+++ src/main/java/com/example/Projeto/student/controller/UpdateStudentsController.java
@@ -1,0 +1,27 @@
+package com.example.Projeto.student;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.*;
+
+import java.time.LocalDate;
+import java.time.Month;
+import java.util.List;
+
+@RestController
+@RequestMapping("api/v1/students")
+public class ListStudentController {
+    private final StudentService studentService;
+
+    @Autowired
+    public StudentController(StudentService studentService) {
+        this.studentService = studentService;
+    }
+
+    @PutMapping(path = "{studentId}")
+    public void updateStudent(@PathVariable("studentId") Long studentId,
+                              @RequestParam(required = false) String name,
+                              @RequestParam(required = false) String email){
+        studentService.updateStudent(studentId, name, email);
+
+    }
+}
